+++ packages/frontend-core/src/utils/index.ts
@@ -16,3 +16,4 @@ export * from "./table"
 export * from "./components"
 export * from "./validation"
 export * from "./formatting"
+export * from "./login"
